move unseparated_literal_suffix to its own module
